student-portal/requirements.txt
flask
gunicorn
mysql-connector-python
werkzeug
